create menu managment done

+++ src/main/java/menuManegment/demo/menu/controller/AbstractCRUDController.java
@@ -1,21 +1,17 @@
 package menuManegment.demo.menu.controller;
 
-import menuManegment.demo.menu.mapper.CatalogMenuManagmentMapper;
 import menuManegment.demo.menu.model.ModelLoadable;
-import menuManegment.demo.menu.repository.CatalogMenuManagementRepository;
 import menuManegment.demo.menu.service.MenuCRUD;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
 
 import static org.springframework.http.ResponseEntity.created;
 
-public abstract class AbstractCRUDController<M extends ModelLoadable<Long>> {
+public abstract class AbstractCRUDController<M extends ModelLoadable<Integer>> {
 
     protected final MenuCRUD<?, M> service;
